Fully implemented TS on container files
